+++ CHANGELOG.md
@@ -25,6 +25,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2024-08-08
+
+### Added
+- Added `get_api_version` method to `CFSClient` that queries the CFS API for the
+  current API version.
+- Added `supports_customized_image_name` property to `CFSClient` that checks
+  whether the CFS version supports specifying the resultant image name for
+  image customization sessions.
+- Added use of the CFS API `image_map` parameter when creating CFS image
+  customization sessions in the `create_image_customization_session` method of
+  the `CFSClient` if the CFS version supports it. This passes the desired name
+  of the customized image to the CFS session.
+
+### Changed
+- Changed the `APIGatewayClient` handling of `base_resource_path` so that it is
+  possible to make a request on the `base_resource_path` without a trailing `/`.
+  This is required to allow the `CFSClient` to do a GET on the endpoint
+  `/cfs/v2`, which returns the semantic version of CFS. If any users of this
+  class were expecting calls to `_make_req` (or `get`, `put`, etc.) without
+  arguments to make a request to `base_resource_path` with a trailing slash,
+  they must be updated to pass an empty string argument. No such instances are
+  known in either the `sat` or `cfs-config-util` users of this library.
+
 ## [1.2.4] - 2024-04-25
 
 ### Fixed